Some more refactoring done.
Added IterationEnumerator to convert to and from Iterators and
Enumerations.
Added StaticURLClassLoaderWrapper, and a ClassLoader Wrapper for
JarUtils (Also seriously, ClassLoaders are a pain in the ass in a
static context)

Fleshed out ThreadUtils a bit
Fleshed out JarUtils a bit

Added the ability for FileUtils to store the programs Assets Directory.

Removed AsyncJarUtil and FutureHandler because screw that noise.